+++ tests/test_scrabble_cli.py
@@ -1,6 +1,7 @@
 import unittest
 from unittest.mock import patch
 from game.scrabble_cli import ScrabbleCli
+from game.square import Square
 from game.tile import Tile
 from io import StringIO
 
@@ -38,7 +39,7 @@ class MyTestCase(unittest.TestCase):
         self.assertEqual(scrabblecli.game.board.grid[7][9].letter, Tile('L', 1))
         self.assertEqual(scrabblecli.game.board.grid[7][10].letter, Tile('A', 1))
         self.assertEqual(len(scrabblecli.game.players[scrabblecli.game.current_player_index].tiles), 0)
-        self.assertEqual(scrabblecli.game.players[scrabblecli.game.current_player_index - 1].score, 8)
+        self.assertEqual(scrabblecli.game.players[scrabblecli.game.current_player_index - 1].score, 10)
 
     @patch('builtins.input', side_effect=['draw', '1'])
     def test_player_draw_cards(self, mock_input):
@@ -98,6 +99,9 @@ class MyTestCase(unittest.TestCase):
     @patch('builtins.input', side_effect=['Dino', 'scores', 'quit'])
     def test_start_game(self, mock_input, mock_stdout):
         scrabblecli = ScrabbleCli()
+        square = Square()
+        square.put_tile(Tile('A', 1))
+        scrabblecli.game.board.grid[7][7] = square
         scrabblecli.game.players.pop(1)
         scrabblecli.game.players[0].set_name("Dino")
         scrabblecli.start_game()
@@ -108,11 +112,71 @@ class MyTestCase(unittest.TestCase):
     @patch('builtins.input', side_effect=['Dino', 'tiles', 'quit'])
     def test_show_tiles(self, mock_input, mock_stdout):
         scrabblecli = ScrabbleCli()
+        square = Square()
+        square.put_tile(Tile('A', 1))
+        scrabblecli.game.board.grid[7][7] = square
         scrabblecli.game.players.pop(1)
         scrabblecli.game.players[0].set_name("Dino")
         scrabblecli.start_game()
         self.assertEqual(mock_stdout.getvalue().strip(), repr(scrabblecli.game.players[0].tiles))
         self.assertEqual(scrabblecli.game_state, "over")
+
+    @patch('sys.stdout', new_callable=StringIO)
+    @patch('builtins.input', side_effect=['Dino', 'not an action', 'quit'])
+    def test_action_not_valid(self, mock_input, mock_stdout):
+        scrabblecli = ScrabbleCli()
+        square = Square()
+        square.put_tile(Tile('A', 1))
+        scrabblecli.game.board.grid[7][7] = square
+        scrabblecli.game.players.pop(1)
+        scrabblecli.start_game()
+        self.assertEqual(mock_stdout.getvalue().strip(), "Action not valid, please choose from: pass, play, draw, "
+                                                         "quit, scores, tiles")
+        self.assertEqual(scrabblecli.game_state, "over")
+
+    @patch('sys.stdout', new_callable=StringIO)
+    @patch('builtins.input', side_effect=['hola', '7', '7', 'horizontal'])
+    def test_first_turn(self, mock_input, mock_stdout):
+        scrabblecli = ScrabbleCli()
+        scrabblecli.game.players.pop(1)
+        scrabblecli.game.players[scrabblecli.game.current_player_index].tiles = \
+            [Tile('H', 1),
+             Tile('O', 1),
+             Tile('L', 1),
+             Tile('A', 1)]
+        scrabblecli.first_turn()
+        self.assertEqual(scrabblecli.game.board.grid[7][7].letter, Tile('H', 1))
+        self.assertEqual(scrabblecli.game.board.grid[7][8].letter, Tile('O', 1))
+        self.assertEqual(scrabblecli.game.board.grid[7][9].letter, Tile('L', 1))
+        self.assertEqual(scrabblecli.game.board.grid[7][10].letter, Tile('A', 1))
+
+    @patch('sys.stdout', new_callable=StringIO)
+    @patch('builtins.input', side_effect=['hola', '0', '7', 'horizontal'])
+    def test_first_turn_wrong(self, mock_input, mock_stdout):
+        scrabblecli = ScrabbleCli()
+        scrabblecli.game.players.pop(1)
+        scrabblecli.game.players[scrabblecli.game.current_player_index].tiles = \
+            [Tile('H', 1),
+             Tile('O', 1),
+             Tile('L', 1),
+             Tile('A', 1)]
+        scrabblecli.first_turn()
+        self.assertFalse(scrabblecli.game.board.grid[0][7].has_tile())
+
+    @patch('sys.stdout', new_callable=StringIO)
+    @patch('builtins.input', side_effect=["Dino", 'hola', '7', '7', 'horizontal', 'quit'])
+    def test_game_start_board_empty(self, mock_input, mock_stdout):
+        scrabblecli = ScrabbleCli()
+        scrabblecli.game.players.pop(1)
+        scrabblecli.game.players[scrabblecli.game.current_player_index].tiles = \
+            [Tile('H', 1),
+             Tile('O', 1),
+             Tile('L', 1),
+             Tile('A', 1)]
+        scrabblecli.start_game()
+        self.assertTrue(scrabblecli.game.board.grid[7][7].has_tile())
+        self.assertEqual(scrabblecli.game.board.grid[7][7].letter, Tile('H', 1))
+
 
 
 if __name__ == '__main__':